+++ src/ErrorsInVariables.jl
@@ -5,17 +5,19 @@ include("estimator.jl")
 include("meive.jl")
 include("orthogonalregression.jl")
 include("deming.jl")
+include("simex.jl")
 
 
 export CGA
 export Estimator
-
+export SimulationExtrapolation
 
 import .Estimator: eive, EiveResult
 import .Eivem: meive
 import .CGA: cga, cgasample, converged 
 import .OrthogonalRegression: orthogonal_regression
 import .DemingRegression: deming
+import .SimulationExtrapolation: simex
 
 
 export eive
@@ -26,6 +28,6 @@ export cgasample
 export converged 
 export orthogonal_regression
 export deming 
-
+export simex
 
 end # module