+++ test/runtests.jl
@@ -7,12 +7,12 @@ function euclidean(v1::Vector, v2::Vector)
 end
 
 include("testdd.jl")
-#include("testiv.jl")
-#include("testcga.jl")
-#include("testestimator.jl")
-#include("testmeive.jl")
-#include("testorthogonalregression.jl")
-#include("testdeming.jl")
-#include("testsimex.jl")
+include("testiv.jl")
+include("testcga.jl")
+include("testestimator.jl")
+include("testmeive.jl")
+include("testorthogonalregression.jl")
+include("testdeming.jl")
+include("testsimex.jl")